Refactor logging of MCP gateways
