+++ ymir/tools/unprivileged/gateway.py
@@ -1,16 +1,12 @@
-import logging
 import os
-import re
-from typing import Any
-import re
 
 from beeai_framework.adapters.mcp.serve.server import (
     MCPServer,
     MCPServerConfig,
     MCPSettings,
 )
-from beeai_framework.emitter.emitter import Emitter
 
+from ymir.tools.gateway_utils import setup_logging
 from ymir.tools.unprivileged.commands import RunShellCommandTool
 from ymir.tools.unprivileged.distgit_detector import DistgitDetectorTool
 from ymir.tools.unprivileged.filesystem import GetCWDTool, RemoveTool
@@ -44,65 +40,6 @@ from ymir.tools.unprivileged.wicked_git import (
     GitPatchCreationTool,
 )
 
-# Patterns that match common credential formats in log output
-_REDACT_PATTERNS = [
-    # GitLab PAT
-    re.compile(r"glpat-[A-Za-z0-9_-]{20,}"),
-    # Anthropic API key
-    re.compile(r"sk-ant-[A-Za-z0-9_-]{20,}"),
-    # Google API key
-    re.compile(r"AIzaSy[A-Za-z0-9_-]{33}"),
-    # Bearer tokens in URLs or strings
-    re.compile(r"oauth2:[^@\s]+@"),
-    # Testing Farm API tokens (UUID format)
-    re.compile(r"[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}", re.IGNORECASE),
-    # Jira Cloud API tokens (ATATT3x... pattern)
-    re.compile(r"ATATT3x[A-Za-z0-9_-]{20,}"),
-    # Base64 Authorization headers
-    re.compile(r"Basic [A-Za-z0-9+/=]{20,}"),
-    # Generic long hex/base64 tokens (e.g. Jira PATs)
-    re.compile(
-        r"(?:token|key|password|secret|credential)[\"'=:\s]+[A-Za-z0-9+/=_-]{20,}['\"\s]*",
-        re.IGNORECASE,
-    ),
-]
-
-logger = logging.getLogger(__name__)
-
-
-def _redact(text: str) -> str:
-    """Replace credential-like patterns in text with [REDACTED]."""
-    for pattern in _REDACT_PATTERNS:
-        text = pattern.sub("[REDACTED]", text)
-    return text
-
-
-def _setup_logging():
-    handlers = [logging.StreamHandler()]
-    if os.environ.get("DEBUG_FILE"):
-        handlers.append(logging.FileHandler(os.environ.get("DEBUG_FILE")))
-    logging.basicConfig(level=logging.INFO, handlers=handlers)
-
-    # Log tool calls via Emitter.
-    # Dotted strings in Emitter.on() are matched exactly (not as globs),
-    # so we use regex patterns to match any tool's events.
-    def on_tool_start(data: Any, meta: Any):
-        logger.info(f"Tool called: {meta.creator}")
-        logger.info(f"Tool arguments: {str(data)}")
-
-    def on_tool_success(data: Any, meta: Any):
-        logger.info(f"Tool {meta.creator} completed successfully")
-
-    def on_tool_error(data: Any, meta: Any):
-        logger.error(f"Tool {meta.creator} failed with error: {str(data)}")
-        error = getattr(data, "error", None)
-        if error is not None:
-            logger.error(f"Tool {meta.creator} traceback:", exc_info=error)
-
-    Emitter.root().on(re.compile(r"^tool\..+\.start$"), on_tool_start)
-    Emitter.root().on(re.compile(r"^tool\..+\.success$"), on_tool_success)
-    Emitter.root().on(re.compile(r"^tool\..+\.error$"), on_tool_error)
-
 
 def main():
     transport = os.getenv("MCP_TRANSPORT", "sse")
@@ -114,7 +51,7 @@ def main():
         )
     config = MCPServerConfig(**config_kwargs)
 
-    _setup_logging()
+    setup_logging()
     mcp = MCPServer(config=config)
     mcp.register_many(
         [
